# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Freelance CI API",
        "description": "API REST de la plateforme Freelance CI — connexion freelances, clients et entrepreneurs en Côte d'Ivoire. Paiements via Genius Pay (Mobile Money, Carte), authentification Supabase JWT.",
        "contact": {
            "name": "Freelance CI",
            "email": "contact@freelance-ci.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "0.6.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement local"
        },
        {
            "url": "https://api.freelance-ci.com/api",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/referrals/code": {
            "get": {
                "tags": [
                    "Referrals"
                ],
                "summary": "Get your referral code",
                "operationId": "d5ff0becb29752014cafb7305acb1978",
                "responses": {
                    "200": {
                        "description": "Referral code",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "code": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/referrals/stats": {
            "get": {
                "tags": [
                    "Referrals"
                ],
                "summary": "Get referral statistics",
                "operationId": "be5ff4fdeb71f175dfc81883055ff150",
                "responses": {
                    "200": {
                        "description": "Referral stats",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "total_referrals": {
                                                    "type": "integer"
                                                },
                                                "completed_referrals": {
                                                    "type": "integer"
                                                },
                                                "total_earned": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "reward_per_referral": {
                                                    "type": "number",
                                                    "format": "float"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/referrals": {
            "get": {
                "tags": [
                    "Referrals"
                ],
                "summary": "List referrals made by you",
                "operationId": "bc2aaf610b4448bf9f1ecc5760b66da4",
                "responses": {
                    "200": {
                        "description": "List of referrals",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Referral"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/contracts/{contract}/review": {
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Create a review for a contract",
                "operationId": "a9b9d60f5687bac282251c4b9bcf49f4",
                "parameters": [
                    {
                        "name": "contract",
                        "in": "path",
                        "description": "Contract ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "rating": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "comment": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Review submitted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Review"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unable to create review",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/freelances/{freelance}/reviews": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "List reviews for a freelance",
                "operationId": "84cdc863a5f248e01e868a11bbbf2087",
                "parameters": [
                    {
                        "name": "freelance",
                        "in": "path",
                        "description": "Freelance user ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of reviews",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Review"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/reviews/{review}/reply": {
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Reply to a review",
                "operationId": "cabbc2999b329b1e1c656af4ebfc2539",
                "parameters": [
                    {
                        "name": "review",
                        "in": "path",
                        "description": "Review ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "comment": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reply submitted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Review"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unable to reply to review",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer"
                    },
                    "last_page": {
                        "type": "integer"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "Review": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "rating": {
                        "type": "integer",
                        "maximum": 5,
                        "minimum": 1
                    },
                    "comment": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Referral": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "referrer_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "referred_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "completed",
                            "paid"
                        ]
                    },
                    "reward_amount": {
                        "type": "number",
                        "format": "float"
                    },
                    "paid_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerToken": {
                "type": "http",
                "description": "JWT Supabase — coller le token reçu après connexion",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "BearerToken": []
        }
    ]
}